% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/belg-package.R
\docType{package}
\name{belg-package}
\alias{belg}
\alias{belg-package}
\title{belg: Boltzmann Entropy of a Landscape Gradient}
\description{
Calculates the Boltzmann entropy of a landscape gradient.
    This package uses the analytical method created by Gao, P., Zhang, H.
    and Li, Z., 2018 (<doi:10.1111/tgis.12315>). It also extend the original 
    idea by allowing calculations on data with missing values.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-spatialecology/belg}
  \item Report bugs at \url{https://github.com/r-spatialecology/belg/issues}
}

}
\author{
\strong{Maintainer}: Jakub Nowosad \email{nowosad.jakub@gmail.com} (0000-0002-1057-3721)

Other contributors:
\itemize{
  \item Space Informatics Lab [copyright holder]
}

}
\keyword{internal}
