% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{complex_land}
\alias{complex_land}
\title{Complex landscape}
\format{An object of class \code{RasterLayer} of dimension 3 x 8 x 1.}
\usage{
complex_land
}
\description{
A dataset containing artificial complex landscape
}
\keyword{datasets}
