% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{land_gradient1}
\alias{land_gradient1}
\title{Complex landscape}
\format{
An object of class \code{RasterLayer} of dimension 512 x 512 x 1.
}
\usage{
land_gradient1
}
\description{
A dataset containing artificial complex landscape
}
\keyword{datasets}
