% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{land_gradient2}
\alias{land_gradient2}
\title{Simple landscape}
\format{
An object of class \code{RasterLayer} of dimension 512 x 512 x 1.
}
\usage{
land_gradient2
}
\description{
A dataset containing artificial simple landscape
}
\keyword{datasets}
