% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{simple_land}
\alias{simple_land}
\title{Simple landscape}
\format{An object of class \code{RasterLayer} of dimension 3 x 8 x 1.}
\usage{
simple_land
}
\description{
A dataset containing artificial simple landscape
}
\keyword{datasets}
