% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{complex_land}
\alias{complex_land}
\title{Complex landscape (small)}
\format{
An object of class \code{RasterLayer} of dimension 6 x 8 x 1.
}
\usage{
complex_land
}
\description{
A dataset containing small artificial complex landscape
}
\keyword{datasets}
