% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_boltzmann}
\alias{get_boltzmann}
\title{Boltzmann entropy of a landscape gradient;}
\usage{
get_boltzmann(x, base = "log", relative = FALSE)
}
\arguments{
\item{x}{A matrix.}

\item{base}{A logarithm base ("log", "log2" or "log10")}

\item{relative}{TRUE/FALSE}
}
\description{
Boltzmann entropy of a landscape gradient;
}
\references{
Gao, Peichao, Hong Zhang, and Zhilin Li. "A hierarchy-based
solution to calculate the configurational entropy of landscape gradients."
Landscape Ecology 32.6 (2017): 1133-1146.

Gao, Peichao, Hong Zhang, and Zhilin Li. "An efficient
analytical method for computing the Boltzmann entropy of a landscape
gradient." Transactions in GIS (2018).
}
